Add GSQuery GameSpy version, update UTQuery
Mostly a drop-in replacement for UTQuery once you update dict keys and team values. GameSpy query doesn't work on all servers and does not include as much information, but it is much faster and more robust. This version includes a bunch of wonky comments, I didn't yet check to see if the codebase is complete. TODO: allow packets to arrive out of order. 

RemoteScoreboard_GS is included as a demo of RemoteScoreboard using the GameSpy